I ran ./configure CPPFLAGS='-Wall --pedantic --W' to generate a bunch more warnings. Cleaned up most, BUT some of the sign compare warnings in the db* code I didn't touch because I was worried I'd break things.  I didn't want to cast if it wasn't clear it was OK.
There are not many left, so it would be helpful to have someone else look
at the output from make when run with the above CPPFLAGS.

I didn't bother "fixing" the C++ style comments.  What compiler can't deal
with those.  I do think the casts and mixed sign compares might get us in
trouble some day if any numbers got very large (e.g. a seek position).

+++ swish-e/src/docprop.c
@@ -792,7 +792,7 @@ propEntry *append_property( struct metaEntry *meta_entry, propEntry *p, char *tx
 
 
     /* Any room to add the property? */
-    if ( meta_entry->max_len &&  p->propLen + add_a_space >=  meta_entry->max_len )
+    if ( meta_entry->max_len &&  (int)p->propLen + add_a_space >=  meta_entry->max_len )
     {
         if ( str )
             efree( str );
